+++ Dogo/src/Dogo/Renderer/Line.h
@@ -34,7 +34,7 @@ namespace Dogo
 				vertexData.push_back(end.y);
 				vertexData.push_back(end.z);
             }
-            m_VertexBuffer.reset(VertexBuffer::Create(m_Vertices.size() * sizeof(assimpVertex), sizeof(assimpVertex), vertexData.data()));
+            m_VertexBuffer.reset(VertexBuffer::Create(uint32_t(m_Vertices.size() * sizeof(assimpVertex)), uint32_t(sizeof(assimpVertex)), vertexData.data()));
             m_VertexBuffer->Bind();
             RenderAPI api = GraphicsContext::Get()->GetAPI();
             switch (api)
